Moved a utility function to utility file
